requirements.txt
face-recognition
python-dotenv